requirements.txt
Django>=5.2.5
djangorestframework>=3.20.0
djangorestframework-simplejwt>=6.0.0
drf-spectacular>=0.28.0
requests>=2.30.0